requirements.txt
Pillow
tkinterdnd2
tk
